backend/app/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-aws
boto3
google-search-results
python-dotenv
slowapi